WIP
I have less hair now.
Two methods show potential.
method the first: Reuse the existing RegionPresentationTargets by having the application send RegionDocumens a message asking for an export. Upon receiving the message, the RegionDocument resizes itself, triggers a render, then saves as png. The problem is that the resizing requires an event: OnRenderSizeChanged, which I can't seem to call or replicate. After sending the message once, the presentation target size is now correct and a second export gets the right output. Since I can't fix the presentation target sizing after changing it (since it need the app to process the message), this causes smaller regions to look wierd after exporting.

The other method is to create a whole new PresentationTarget and set its size. This would hopefully have the benefit of not needing to be hooked into the game's loop. This seems to work, I get an output image of the right size, but it's not getting terrain rendered to it and is just a transparent rectangle.

If I could either:
Process the OnRenderSizeChanged events that I'm queuing up; or figure out how to get a GraphicsScreen to render on demand, I could get clean maps.