Initial Feature commit
Right click a selection, or hit Ctrl-R while in a world render to summon the Replace... dialog. Choose to either replace all of a staticID across all components in the selection (useful for changing floor or wall statics) or to replace all of a kind of component with another (useful for turning all the statics to floors or trees to walls).

Still to do, handle things like asking if the new walls should be indestructible, or how much damage the whirlwind or poison water should do.